+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 22.12.0
 NEW FEATURES:
 * resource/cvo_volume: support create and delete onPrem volume.
+* resource/cvo_volume: support create snapshot policy for AWS, AZURE and GCP if the snapshot policy is not available.
 
 ## 22.11.0
 NEW FEATURES: